mm: vmpressure: Fix a race that would erroneously clear accumulated data
Since the code that determines whether data should be cleared and the
code that actually clears the data are in separate spin-locked critical
sections, new data could be generated on another CPU after it is
determined that the existing data should be cleared, but before the
current CPU clears the existing data. This would cause the new data
reported by the other CPU to be lost.

Fix the race by clearing accumulated data within the same spin-locked
critical section that determines whether or not data should be cleared.

Change-Id: Id221878ce1856fd6635240490b137f76bf0a4491
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>